+++ civis/cli/_cli_commands.py
@@ -3,7 +3,7 @@
 """
 Additional commands to add to the CLI beyond the OpenAPI spec.
 """
-
+import functools
 import os
 
 import click
@@ -68,6 +68,83 @@ def files_download_cmd(file_id, path):
     """Download a Civis File to a specified local path."""
     with open(path, 'wb') as f:
         civis_to_file(file_id, f)
+
+
+@click.command('download')
+@click.argument('database_name', type=str)
+@click.argument('file_name', type=click.Path(exists=True))
+@click.argument('output_file', type=click.Path())
+def sql_download_cmd(database_name, file_name, output_file):
+    """Download results of a query."""
+    with open(file_name, 'rt') as f:
+        sql = f.read()
+    fut = civis.io.civis_to_csv(output_file, sql, database=database_name)
+    fut.result()
+    print("Downloaded the result of the query to %s." % output_file)
+
+
+@click.command('run')
+@click.argument('database_name', type=str)
+@click.argument('file_name', type=click.Path(exists=True))
+@click.option('-n', type=int, default=100,
+              help="Display up to this many rows of the result. Max 100.")
+def sql_run_cmd(database_name, file_name, n):
+    """Run the SQL in the specified file."""
+    with open(file_name, 'rt') as f:
+        sql_cmd = f.read()
+    print('\nExecuting query...')
+    fut = civis.io.query_civis(sql_cmd, database=database_name, preview_rows=n,
+                               polling_interval=3)
+    cols, rows = fut.result()['result_columns'], fut.result()['result_rows']
+    print('...Query complete.\n')
+    print(_str_table_result(cols, rows))
+
+
+@click.command('cmd')
+@click.argument('database_name', type=str)
+@click.argument('sql_cmd', type=str, default=None, required=False)
+@click.option('-n', type=int, default=100,
+              help="Display up to this many rows of the result. Max 100.")
+def sql_cmd_cmd(database_name, sql_cmd, n):
+    """Run a SQL command."""
+    if sql_cmd is None:
+        # Read the SQL query from user input. This also allows use of a heredoc
+        lines = []
+        while True:
+            try:
+                _i = input()
+            except (KeyboardInterrupt, EOFError):
+                # The end of a heredoc produces an EOFError.
+                break
+            if not _i:
+                break
+            else:
+                lines.append(_i)
+        sql_cmd = '\n'.join(lines)
+    print('\nExecuting query...')
+    fut = civis.io.query_civis(sql_cmd, database=database_name, preview_rows=n,
+                               polling_interval=3)
+    cols, rows = fut.result()['result_columns'], fut.result()['result_rows']
+    print('...Query complete.\n')
+    print(_str_table_result(cols, rows))
+
+
+def _str_table_result(cols, rows):
+    """Turn a Civis Query result into a readable table."""
+    # Determine the maximum width of each column.
+    # First find the width of each element in each row, then find the max
+    # width in each position.
+    max_len = functools.reduce(
+        lambda x, y: [max(z) for z in zip(x, y)],
+        [[len(_v) for _v in _r] for _r in [cols] + rows])
+
+    header_str = " | ".join("{0:<{width}}".format(_v, width=_l)
+                            for _l, _v in zip(max_len, cols))
+    tb_strs = [header_str, len(header_str) * '-']
+    for row in rows:
+        tb_strs.append(" | ".join("{0:>{width}}".format(_v, width=_l)
+                                  for _l, _v in zip(max_len, row)))
+    return '\n'.join(tb_strs)
 
 
 @click.command('download')